+++ internal/backup/backup.go
@@ -22,6 +22,7 @@ type Backup struct {
 	remotePath   string
 	localDir     string
 	dumpLocation string
+	removeDump   bool
 }
 
 func NewApp(
@@ -31,6 +32,7 @@ func NewApp(
 	remotePath,
 	localDir,
 	dumpLocation string,
+	removeDump bool,
 ) *Backup {
 	return &Backup{
 		ctx:          ctx,
@@ -39,6 +41,7 @@ func NewApp(
 		remotePath:   remotePath,
 		localDir:     localDir,
 		dumpLocation: dumpLocation,
+		removeDump:   removeDump,
 	}
 }
 
@@ -61,7 +64,7 @@ func (b *Backup) Backup() error {
 
 func (b *Backup) backupByServer() error {
 
-	isRemoveDump := true
+	isRemoveDump := b.removeDump
 	checkCmd := fmt.Sprintf("test -f %s", b.remotePath)
 
 	logging.L(b.ctx).Info(
@@ -82,8 +85,8 @@ func (b *Backup) backupByServer() error {
 		stop := make(chan struct{})
 		dumpCreateTimeNow := time.Now()
 
-		logging.L(b.ctx).Info("File dump", logging.StringAttr("name", b.remotePath))
-		fmt.Println("File dump:", b.remotePath)
+		logging.L(b.ctx).Info("File dump name", logging.StringAttr("name", b.remotePath))
+		fmt.Println("File dump name:", b.remotePath)
 
 		go utils.Spinner(stop)
 
@@ -99,12 +102,20 @@ func (b *Backup) backupByServer() error {
 		close(stop)
 
 		elapsed := time.Since(dumpCreateTimeNow)
+
+		totalSize, err := b.fileSize()
+		if err != nil {
+			return err
+		}
+
 		fmt.Printf("\rDump created successfully in %.2f sec\n", elapsed.Seconds())
+		fmt.Printf("\rFile dump size: %s [%d bytes]\n", utils.FormatBytes(totalSize), totalSize)
 
 		dumpCreateTimeSec := fmt.Sprintf("%.2f sec", elapsed.Seconds())
 		logging.L(b.ctx).Info(
 			"The dump was successfully created",
 			logging.StringAttr("time", dumpCreateTimeSec),
+			logging.Int64Attr("size", totalSize),
 		)
 	}
 
@@ -147,15 +158,9 @@ func (b *Backup) backupLocalDirect() error {
 func (b *Backup) downloadFile() error {
 	localPath := filepath.Join(b.localDir, filepath.Base(b.remotePath))
 
-	sizeOutput, err := b.conn.RunCommand(fmt.Sprintf("stat -c %%s %s", b.remotePath))
-	if err != nil {
-		return fmt.Errorf("failed to get file size: %v", err)
-	}
-	sizeOutput = strings.TrimSpace(sizeOutput)
-
 	var totalSize int64
 
-	_, err = fmt.Sscanf(sizeOutput, "%d", &totalSize)
+	totalSize, err := b.fileSize()
 	if err != nil {
 		return err
 	}
@@ -211,4 +216,22 @@ func (b *Backup) downloadFile() error {
 	fmt.Println("\nDownload complete:", localPath)
 
 	return session.Wait()
+}
+
+func (b *Backup) fileSize() (int64, error) {
+	sizeOutput, err := b.conn.RunCommand(fmt.Sprintf("stat -c %%s %s", b.remotePath))
+
+	var totalSize int64
+
+	if err != nil {
+		return totalSize, fmt.Errorf("failed to get file size: %v", err)
+	}
+	sizeOutput = strings.TrimSpace(sizeOutput)
+
+	_, err = fmt.Sscanf(sizeOutput, "%d", &totalSize)
+	if err != nil {
+		return totalSize, err
+	}
+
+	return totalSize, nil
 }